optimize: accelerate matrix multiplication with SSE instructions (fixed several test blockers BTW)
